requirements.txt
langchain>=0.2.0
langchain-community>=0.0.195
# langchain-huggingface>=0.0.3
faiss-cpu==1.7.4
sentence-transformers>=2.2.2
pypdf>=3.11.0
streamlit>=1.25.0
numpy>=1.24.0,<2.0
openai>=1.30.0
python-dotenv>=1.0.0
